Add  remove cycle mergeRecords -> incomingSyncUtils -> mergeRecords
